convert to typescript and fixed development and build and start and more ...
