+++ package.json
@@ -4,8 +4,16 @@
   "description": "# An MVC framework written in JS",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "serve": "node server.js"
+    "build": "npx tsc",
+    "start": "node dist/index.js",
+    "dev": "nodemon server.ts"
+  },
+  "nodemonConfig": {
+    "watch": [
+      "**"
+    ],
+    "ext": "ts",
+    "exec": "ts-node server.ts"
   },
   "keywords": [],
   "author": "Muhammadreza Haghiri",
@@ -14,5 +22,12 @@
     "express": "^4.18.2",
     "sequelize": "^6.37.1",
     "sqlite3": "^5.1.7"
+  },
+  "devDependencies": {
+    "@types/express": "^4.17.21",
+    "@types/node": "^20.11.24",
+    "nodemon": "^3.1.0",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.3.3"
   }
 }
