+++ routes/index.ts
@@ -1,5 +1,5 @@
-const express = require('express');
-const SampleController = require('../controllers/SampleController');
+import express from 'express';
+import SampleController from '../controllers/SampleController';
 const router = express.Router();
 
 /*
@@ -16,6 +16,7 @@ It'll be more automated as soon as possible.
 
 // Define routes
 
-router.get('/sample', SampleController.getIndex)
+router.get('/sample', SampleController.getIndex);
 
-module.exports = router;
+const executeRoutes = router;
+export default executeRoutes;
